+++ src/main/java/com/moving_mgmt/controller/ApartmentController.java
@@ -7,11 +7,14 @@ import com.moving_mgmt.repositories.AmenitiesRepository;
 import com.moving_mgmt.repositories.ApartmentRepository;
 import com.moving_mgmt.repositories.LocationRepository;
 import com.moving_mgmt.repositories.RoomsRepository;
+import com.moving_mgmt.utils.JwtAuthentication;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -39,28 +42,40 @@ public class ApartmentController {
 
     // add an apartment
     @PostMapping("/api/apartments")
-    public Apartment addApartment(@RequestBody Apartment apartment) {
-        repository.save(apartment);
-        return apartment;
+    public Apartment addApartment(@RequestBody Apartment apartment, HttpServletRequest request) throws Exception {
+        String token = request.getHeader(HttpHeaders.AUTHORIZATION);
+        if(!JwtAuthentication.decodeJWT(token).equals(IllegalArgumentException.class)) {
+            repository.save(apartment);
+            return apartment;
+        }
+        throw new IllegalAccessException("invalid token");
     }
 
     // update an apartment
     @PutMapping("/api/apartments/{id}")
-    public Apartment updateApartment(@PathVariable Integer id, @RequestBody Apartment apartment) throws Exception {
-        Apartment tempApartment = repository.findApartmentById(id);
+    public Apartment updateApartment(@PathVariable Integer id, @RequestBody Apartment apartment, HttpServletRequest request) throws Exception {
+        String token = request.getHeader(HttpHeaders.AUTHORIZATION);
+        if(!JwtAuthentication.decodeJWT(token).equals(IllegalArgumentException.class)) {
+            Apartment tempApartment = repository.findApartmentById(id);
 
-        if(!tempApartment.equals(null)) {
-            apartment.setId(tempApartment.getId());
-            repository.save(apartment);
+            if(!tempApartment.equals(null)) {
+                apartment.setId(tempApartment.getId());
+                repository.save(apartment);
+            }
+            return apartment;
         }
-        return apartment;
+        throw new IllegalAccessException("invalid token");
     }
 
     // remove an apartment
     @DeleteMapping("/api/apartments/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @Transactional
-    public void deleteApartment(@PathVariable Integer id) {
-        repository.deleteById(id);
+    public void deleteApartment(@PathVariable Integer id, HttpServletRequest request) throws Exception {
+        String token = request.getHeader(HttpHeaders.AUTHORIZATION);
+        if(!JwtAuthentication.decodeJWT(token).equals(IllegalArgumentException.class)) {
+            repository.deleteById(id);
+        }
+        throw new IllegalAccessException("invalid token");
     }
 }
